consolidate code and move helpers to separate file
