+++ matchms/similarity/FlashSpectralEntropy.py
@@ -7,6 +7,11 @@ from sparsestack import StackedSparseArray
 from tqdm import tqdm
 from matchms.typing import SpectrumType
 from .BaseSimilarity import BaseSimilarity
+from .flash_utils import (
+    _build_library_index,
+    _clean_and_weight,
+    _LibraryIndex,
+)
 
 
 class FlashSpectralEntropy(BaseSimilarity):
@@ -98,6 +103,7 @@ class FlashSpectralEntropy(BaseSimilarity):
                                     noise_cutoff=self.noise_cutoff,
                                     normalize_to_half=self.normalize_to_half,
                                     merge_within_da=self.merge_within,
+                                    score_type="entropy",
                                     dtype=self.dtype)
         return cleaned, (None if pmz is None else float(pmz))
 
@@ -248,12 +254,13 @@ class FlashSpectralEntropy(BaseSimilarity):
                                         noise_cutoff=self.noise_cutoff,
                                         normalize_to_half=self.normalize_to_half,
                                         merge_within_da=self.merge_within,
+                                        score_type="entropy",
                                         dtype=self.dtype)
             lib_proc.append(cleaned)
             lib_pmz.append(None if pmz is None else float(pmz))
 
-        build_nl = (self.mode in ("neutral_loss", "hybrid"))
-        lib = _build_library_index(lib_proc, lib_pmz, build_neutral_loss=build_nl, dtype=self.dtype)
+        compute_nl = (self.mode in ("neutral_loss", "hybrid"))
+        lib = _build_library_index(lib_proc, lib_pmz, compute_neutral_loss=compute_nl, dtype=self.dtype)
 
         # 2) prepare output
         out = np.zeros((n_rows, n_cols), dtype=self.dtype)
@@ -269,7 +276,7 @@ class FlashSpectralEntropy(BaseSimilarity):
             tol=float(self.tolerance),
             use_ppm=bool(self.use_ppm),
             mode=self.mode,
-            build_nl=build_nl,
+            compute_nl=compute_nl,
             iden_tol=(None if self.identity_precursor_tolerance is None else float(self.identity_precursor_tolerance)),
             iden_use_ppm=bool(self.identity_use_ppm),
         )
@@ -338,9 +345,6 @@ class FlashSpectralEntropy(BaseSimilarity):
 
 # ===================== helpers =====================
 
-def _as_dtype(a: np.ndarray, dtype: np.dtype) -> np.ndarray:
-    return a.astype(dtype, copy=False) if a.dtype == dtype else a.astype(dtype, copy=True)
-
 def _xlog2_vec(x: np.ndarray, dtype: np.dtype) -> np.ndarray:
     # stable x*log2(x) with x=0 -> 0, in requested dtype (float32 default)
     out = np.zeros_like(x, dtype=dtype)
@@ -353,260 +357,6 @@ def _xlog2_scalar(x: float, dtype: np.dtype) -> float:
     if x <= 0.0:
         return 0.0
     return x * np.log2(x)
-
-def _within_tol(m1: float, m2: float, tol: float, use_ppm: bool, dtype: np.dtype) -> bool:
-    if not use_ppm:
-        return abs(m1 - m2) <= tol
-    return abs(m1 - m2) <= tol * 1e-6 * (0.5 * (m1 + m2))
-
-def _search_window_halfwidth(m: float, tol: float, use_ppm: bool, dtype: np.dtype) -> float:
-    if not use_ppm:
-        return tol
-    c = tol * 1e-6
-    denom = 1.0 - 0.5 * c
-    return (c * m) / denom if denom > 0 else (c * m * 2.0)
-
-
-# ===================== preprocessing (Flash rules) =====================
-
-def _entropy_weight(intensities: np.ndarray, dtype: np.dtype) -> np.ndarray:
-    intensities = _as_dtype(np.maximum(intensities, 0.0), dtype)
-    total = float(intensities.sum(dtype=np.float64))  # sum in high precision for stability
-    if total <= 0.0:
-        return intensities
-    p = _as_dtype(intensities / total, dtype)
-    with np.errstate(divide="ignore", invalid="ignore"):
-        logp = np.zeros_like(p, dtype=dtype)
-        mask = p > 0
-        logp[mask] = np.log2(p[mask]).astype(dtype, copy=False)
-    # entropy in bits
-    S = float((-p * logp).sum(dtype=np.float64))
-    w = 1.0 if S >= 3.0 else (0.25 + 0.25 * S)
-    return np.power(intensities, w).astype(dtype, copy=False)
-
-def _merge_within(peaks: np.ndarray, max_delta_da: float, dtype: np.dtype) -> np.ndarray:
-    if peaks.shape[0] <= 1 or max_delta_da <= 0.0:
-        return _as_dtype(peaks, dtype)
-    mz = peaks[:, 0].astype(dtype, copy=True)
-    intensities = peaks[:, 1].astype(dtype, copy=True)
-    new_mz = []
-    new_int = []
-    current_mz = mz[0]
-    current_int = intensities[0]
-    for k in range(1, mz.shape[0]):
-        if (mz[k] - current_mz) <= max_delta_da:
-            total = current_int + intensities[k]
-            if total > 0:
-                current_mz = (current_mz * current_int + mz[k] * intensities[k]) / total
-                current_int = total
-            else:
-                current_mz = mz[k]
-                current_int = 0.0
-        else:
-            new_mz.append(float(current_mz))
-            new_int.append(float(current_int))
-            current_mz = mz[k]
-            current_int = intensities[k]
-    new_mz.append(float(current_mz))
-    new_int.append(float(current_int))
-    out = np.column_stack((np.array(new_mz, dtype=dtype), np.array(new_int, dtype=dtype)))
-    return out
-
-def _clean_and_weight(peaks: np.ndarray,
-                      precursor_mz: Optional[float],
-                      remove_precursor: bool,
-                      precursor_window: float,
-                      noise_cutoff: float,
-                      normalize_to_half: bool,
-                      merge_within_da: float,
-                      dtype: np.dtype) -> np.ndarray:
-    """
-    Apply the Flash preprocessing rules to a (mz, intensity) peak list.
-
-    Steps:
-      1) (Optional) Remove all peaks at/above (precursor_mz - precursor_window).
-      2) (Optional) Remove noise: keep peaks with intensity >= noise_cutoff * max(intensity).
-      3) Entropy-weight intensities (Li & Fiehn): raise intensities by a power derived from spectrum entropy.
-      4) (Optional) Merge peaks within a small m/z window by intensity-weighted centroid.
-      5) Sort by m/z and (optional) normalize intensities to sum to 0.5 (recommended in the paper).
-
-    Parameters
-    ----------
-    peaks : ndarray of shape (N, 2)
-        Column 0 = m/z, column 1 = intensity.
-    precursor_mz : float or None
-        Precursor m/z for the spectrum (if known).
-    remove_precursor : bool
-        If True, remove the precursor and near-precursor peaks.
-    precursor_window : float
-        Size (Da) to cut below the precursor (remove m/z > precursor_mz - window).
-    noise_cutoff : float
-        Fraction of max intensity to keep (0.01 means keep peaks >= 1% of max).
-    normalize_to_half : bool
-        If True, scale intensities so their sum is 0.5.
-    merge_within_da : float
-        If > 0, merge peaks that are within this m/z distance.
-    dtype : np.dtype
-        Float dtype for outputs, usually np.float32.
-
-    Returns
-    -------
-    ndarray
-        Cleaned array with columns [mz, intensity], dtype=dtype, sorted by m/z.
-        May be empty with shape (0, 2) if everything is filtered away.
-    """
-    if peaks.size == 0:
-        return np.empty((0, 2), dtype=dtype)
-
-    mz = _as_dtype(peaks[:, 0], dtype)
-    intensities = _as_dtype(peaks[:, 1], dtype)
-
-    if remove_precursor and (precursor_mz is not None):
-        mask = mz <= (precursor_mz - precursor_window)
-        mz, intensities = mz[mask], intensities[mask]
-        if mz.size == 0:
-            return np.empty((0, 2), dtype=dtype)
-
-    if noise_cutoff and noise_cutoff > 0.0:
-        #thr = dtype.type(mz.dtype.type(inten.max()) * noise_cutoff)
-        thr = intensities.max() * noise_cutoff
-        mask = intensities >= thr
-        mz, intensities = mz[mask], intensities[mask]
-        if mz.size == 0:
-            return np.empty((0, 2), dtype=dtype)
-
-    intensities = _entropy_weight(intensities, dtype)
-
-    if merge_within_da and merge_within_da > 0.0 and mz.size > 1:
-        peaks = _merge_within(np.column_stack((mz, intensities)), merge_within_da, dtype)
-        mz, intensities = peaks[:, 0], peaks[:, 1]
-    else:
-        order = np.argsort(mz)
-        mz, intensities = mz[order], intensities[order]
-
-    s = float(intensities.sum(dtype=np.float64))
-    if s > 0.0 and normalize_to_half:
-        intensities = (intensities * (0.5 / s)).astype(dtype, copy=False)
-
-    return np.column_stack((mz, intensities))
-
-
-# ===================== library index =====================
-
-class _LibraryIndex:
-    """
-    Compact container for the concatenated (sorted) library peaks and, optionally,
-    the neutral-loss view. All arrays are read-only in workers.
-    """
-    __slots__ = (
-        "n_specs",
-        "peaks_mz", "peaks_int", "peaks_spec_idx",
-        "nl_mz", "nl_int", "nl_spec_idx", "nl_product_idx",
-        "precursor_mz",
-        "dtype",
-    )
-    def __init__(self, dtype: np.dtype):
-        self.n_specs = 0
-        self.peaks_mz = None
-        self.peaks_int = None
-        self.peaks_spec_idx = None
-        self.nl_mz = None
-        self.nl_int = None
-        self.nl_spec_idx = None
-        self.nl_product_idx = None
-        self.precursor_mz = None
-        self.dtype = dtype
-
-def _build_library_index(processed_peaks_list: List[np.ndarray],
-                         precursor_mz_list: List[Optional[float]],
-                         build_neutral_loss: bool,
-                         dtype: np.dtype) -> _LibraryIndex:
-    """
-    Build a global, sorted index over all *query* spectra peaks.
-
-    The index concatenates all query peaks into flat arrays, then sorts by m/z.
-    This allows each reference spectrum (a row) to scan efficiently using
-    binary searches into the shared arrays. If `build_neutral_loss` is True, we
-    also construct a parallel set of arrays for neutral-loss peaks (precursor - mz).
-
-    Arrays created (all aligned/sorted):
-      - peaks_mz : float[dtype], all concatenated product (fragment) m/z
-      - peaks_int : float[dtype], matching intensities
-      - peaks_spec_idx : int32, which query spectrum each peak originated from
-      - nl_mz, nl_int, nl_spec_idx, nl_product_idx (optional):
-          neutral-loss m/z for peaks with known precursor,
-          the intensities/spec_idx aligned to nl_mz,
-          and `nl_product_idx` maps back into peaks_mz positions (for hybrid rules)
-
-    Returns
-    -------
-    _LibraryIndex
-        Compact structure used by the row workers to accumulate scores quickly.
-    """
-    idx = _LibraryIndex(dtype)
-    idx.n_specs = len(processed_peaks_list)
-    # store precursor m/z as float32 (NaN when unknown)
-    prec = np.full(idx.n_specs, np.nan, dtype=dtype)
-    for k, pmz in enumerate(precursor_mz_list):
-        if pmz is not None:
-            prec[k] = pmz
-    idx.precursor_mz = prec
-
-    counts = [p.shape[0] for p in processed_peaks_list]
-    N = int(np.sum(counts))
-    if N == 0:
-        idx.peaks_mz = np.zeros(0, dtype=dtype)
-        idx.peaks_int = np.zeros(0, dtype=dtype)
-        idx.peaks_spec_idx = np.zeros(0, dtype=np.int32)
-        if build_neutral_loss:
-            idx.nl_mz = np.zeros(0, dtype=dtype)
-            idx.nl_int = np.zeros(0, dtype=dtype)
-            idx.nl_spec_idx = np.zeros(0, dtype=np.int32)
-            idx.nl_product_idx = np.zeros(0, dtype=np.int64)
-        return idx
-
-    mz_flat = np.empty(N, dtype=dtype)
-    int_flat = np.empty(N, dtype=dtype)
-    spec_flat = np.empty(N, dtype=np.int32)
-    write = 0
-    for s_i, p in enumerate(processed_peaks_list):
-        n = p.shape[0]
-        if n == 0:
-            continue
-        mz_flat[write:write+n] = p[:, 0]
-        int_flat[write:write+n] = p[:, 1]
-        spec_flat[write:write+n] = s_i
-        write += n
-
-    order = np.argsort(mz_flat)
-    idx.peaks_mz = mz_flat[order]
-    idx.peaks_int = int_flat[order]
-    idx.peaks_spec_idx = spec_flat[order]
-    product_pos = np.empty(N, dtype=np.int64)
-    product_pos[order] = np.arange(N, dtype=np.int64)
-
-    if build_neutral_loss:
-        pmz_per_peak = idx.precursor_mz[spec_flat]  # float32, NaN if unknown
-        have_pmz = ~np.isnan(pmz_per_peak)
-        src_idx = np.nonzero(have_pmz)[0]
-        if src_idx.size == 0:
-            idx.nl_mz = np.zeros(0, dtype=dtype)
-            idx.nl_int = np.zeros(0, dtype=dtype)
-            idx.nl_spec_idx = np.zeros(0, dtype=np.int32)
-            idx.nl_product_idx = np.zeros(0, dtype=np.int64)
-        else:
-            nl_mz = (pmz_per_peak[src_idx] - mz_flat[src_idx]).astype(dtype, copy=False)
-            nl_int = int_flat[src_idx]
-            nl_spec = spec_flat[src_idx]
-            nl_prod = product_pos[src_idx]
-
-            order_nl = np.argsort(nl_mz)
-            idx.nl_mz = nl_mz[order_nl]
-            idx.nl_int = nl_int[order_nl]
-            idx.nl_spec_idx = nl_spec[order_nl]
-            idx.nl_product_idx = nl_prod[order_nl]
-
-    return idx
 
 
 # --- Numba-accelerated accumulators ------------------------------------------
@@ -843,7 +593,7 @@ def _row_task_dense(args):
                                    float(cfg["tol"]), bool(cfg["use_ppm"]))
 
     # neutral-loss / hybrid ONLY if library has NL view
-    if cfg["build_nl"]:
+    if cfg["compute_nl"]:
         prefer_frag = (cfg["mode"] == "hybrid")
         if prefer_frag:
             prod_min = np.empty(q_arr.shape[0], dtype=np.int64)
